Add parameter `revisionID` to payload of /notify
Right now the only information sent by wit to notification service is
workitem-id. Based on that generating effective notification is hard
specially when there are updates.

So this commit adds a parameter to the endpoint /notify called
`revisionID` which wit can use to send the UUID of the
event/notification. With this information notification can be generated
correctly.